requirements.txt
Flask==1.1.2
jsonlines==3.1.0
langchain==0.0.171
openai==0.27.4
pinecone_client==2.2.1
python-dotenv==1.0.0
Jinja2==3.0.3
gunicorn==20.1.0
werkzeug==2.0.3
itsdangerous==2.0.1
openai==0.27.4
Chroma==0.2.0
chromadb==0.3.23
pdf2image==1.16.3
pytesseract==0.3.10
tiktoken==0.4.0
unstructured==0.6.6